# engineering/contracts/http/openapi.json
{
  "components": {
    "schemas": {
...
      "CorrelationId": {
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
        "type": "string"
      },
      "CreatePropertyRequest": {
        "additionalProperties": false,
        "properties": {
          "apartmentSubtype": {
            "enum": [
              "MULTI_ROOM",
              "STUDIO"
            ],
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "type": "string"
          },
          "location": {
            "additionalProperties": false,
            "properties": {
              "addressLine": {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              "city": {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              "country": {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              "district": {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              }
            },
            "required": [
              "addressLine",
              "city",
              "country",
              "district"
            ],
            "type": "object"
          },
          "propertyType": {
            "enum": [
              "APARTMENT",
              "COMMERCIAL",
              "HOUSE",
              "LAND",
              "OTHER"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "transactionType": {
            "enum": [
              "LONG_TERM_RENTAL",
              "SALE",
              "SHORT_TERM_RENTAL"
            ],
            "type": "string"
          }
        },
        "required": [
          "location",
          "propertyType",
          "title",
          "transactionType"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "$ref": "#/components/schemas/CorrelationId"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemDetailsError"
            },
            "type": "array"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "code",
          "correlationId",
          "status",
          "title",
          "type"
        ],
        "type": "object"
      },
      "ProblemDetailsError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "path"
        ],
        "type": "object"
      },
...
      "PropertyId": {
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
        "type": "string"
      },
...
      "PropertyPhoto": {
        "additionalProperties": false,
        "properties": {
          "availableAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "category": {
            "enum": [
              "BATHROOM_OR_SHOWER_ROOM",
              "BEDROOM_OR_SLEEPING_AREA",
              "BUILDING_EXTERIOR_OR_ENTRANCE",
              "KITCHEN_OR_KITCHENETTE",
              "LIVING_ROOM_OR_MAIN_ROOM",
              "MAIN_LIVING_SLEEPING_AREA",
              "OTHER"
            ],
            "type": "string"
          },
          "contentByteSize": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "contentPath": {
            "pattern": "^\\/v1\\/properties\\/[0-9a-f-]+\\/photos\\/[0-9a-f-]+\\/content$",
            "type": "string"
          },
          "contentSha256": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          },
          "contentType": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "photoId": {
            "$ref": "#/components/schemas/PropertyId"
          },
          "registeredAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "status": {
            "const": "AVAILABLE",
            "type": "string"
          }
        },
        "required": [
          "availableAt",
          "category",
          "contentByteSize",
          "contentPath",
          "contentSha256",
          "contentType",
          "isPrimary",
          "photoId",
          "registeredAt",
          "status"
        ],
        "type": "object"
      },
...
      "PropertyPortfolioResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PropertyPortfolioItem"
            },
            "type": "array"
          },
          "pageInfo": {
            "additionalProperties": false,
            "properties": {
              "hasNextPage": {
                "type": "boolean"
              },
              "nextCursor": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "hasNextPage",
              "nextCursor"
            ],
            "type": "object"
          }
        },
        "required": [
          "items",
          "pageInfo"
        ],
        "type": "object"
      },
      "PropertyResponse": {
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "apartmentSubtype": {
                "enum": [
                  "MULTI_ROOM",
                  "STUDIO"
                ],
                "type": "string"
              },
              "commercialTerms": {
                "oneOf": [
                  {
                    "additionalProperties": false,
                    "properties": {
                      "chargesAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "currency": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "kind": {
                        "const": "LONG_TERM_RENTAL",
                        "type": "string"
                      },
                      "rentAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "rentPeriod": {
                        "const": "MONTH",
                        "type": "string"
                      },
                      "securityDepositAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "currency",
                      "kind",
                      "rentAmountMinor",
                      "rentPeriod"
                    ],
                    "type": "object"
                  },
                  {
                    "additionalProperties": false,
                    "properties": {
                      "currency": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "kind": {
                        "const": "SHORT_TERM_RENTAL",
                        "type": "string"
                      },
                      "pricingUnit": {
                        "enum": [
                          "NIGHT",
                          "WEEK"
                        ],
                        "type": "string"
                      },
                      "rateAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "currency",
                      "kind",
                      "pricingUnit",
                      "rateAmountMinor"
                    ],
                    "type": "object"
                  },
                  {
                    "additionalProperties": false,
                    "properties": {
                      "currency": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "kind": {
                        "const": "SALE",
                        "type": "string"
                      },
                      "salePriceAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "currency",
                      "kind",
                      "salePriceAmountMinor"
                    ],
                    "type": "object"
                  }
                ]
              },
              "createdAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "details": {
                "additionalProperties": false,
                "properties": {
                  "bathrooms": {
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "bedrooms": {
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "furnished": {
                    "type": "boolean"
                  },
                  "rooms": {
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "usableSurfaceSquareMeters": {
                    "exclusiveMinimum": 0,
                    "type": "number"
                  }
                },
                "type": "object"
              },
              "location": {
                "additionalProperties": false,
                "properties": {
                  "addressLine": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "city": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "country": {
                    "pattern": "^[A-Z]{2}$",
                    "type": "string"
                  },
                  "district": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "addressLine",
                  "city",
                  "country",
                  "district"
                ],
                "type": "object"
              },
              "photos": {
                "items": {
                  "$ref": "#/components/schemas/PropertyPhoto"
                },
                "type": "array"
              },
              "primaryPhoto": {
                "$ref": "#/components/schemas/PropertyPhoto"
              },
              "propertyId": {
                "$ref": "#/components/schemas/PropertyId"
              },
              "propertyType": {
                "enum": [
                  "APARTMENT",
                  "COMMERCIAL",
                  "HOUSE",
                  "LAND",
                  "OTHER"
                ],
                "type": "string"
              },
              "status": {
                "const": "DRAFT",
                "type": "string"
              },
              "structuralRole": {
                "enum": [
                  "COMPOSITE",
                  "STANDALONE",
                  "UNIT"
                ],
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "transactionType": {
                "enum": [
                  "LONG_TERM_RENTAL",
                  "SALE",
                  "SHORT_TERM_RENTAL"
                ],
                "type": "string"
              },
              "updatedAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "type": "string"
              }
            },
            "required": [
              "createdAt",
              "location",
              "photos",
              "propertyId",
              "propertyType",
              "status",
              "structuralRole",
              "title",
              "transactionType",
              "updatedAt"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "apartmentSubtype": {
                "enum": [
                  "MULTI_ROOM",
                  "STUDIO"
                ],
                "type": "string"
              },
              "commercialTerms": {
                "oneOf": [
                  {
                    "additionalProperties": false,
                    "properties": {
                      "chargesAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "currency": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "kind": {
                        "const": "LONG_TERM_RENTAL",
                        "type": "string"
                      },
                      "rentAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "rentPeriod": {
                        "const": "MONTH",
                        "type": "string"
                      },
                      "securityDepositAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "currency",
                      "kind",
                      "rentAmountMinor",
                      "rentPeriod"
                    ],
                    "type": "object"
                  },
                  {
                    "additionalProperties": false,
                    "properties": {
                      "currency": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "kind": {
                        "const": "SHORT_TERM_RENTAL",
                        "type": "string"
                      },
                      "pricingUnit": {
                        "enum": [
                          "NIGHT",
                          "WEEK"
                        ],
                        "type": "string"
                      },
                      "rateAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "currency",
                      "kind",
                      "pricingUnit",
                      "rateAmountMinor"
                    ],
                    "type": "object"
                  },
                  {
                    "additionalProperties": false,
                    "properties": {
                      "currency": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "kind": {
                        "const": "SALE",
                        "type": "string"
                      },
                      "salePriceAmountMinor": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "currency",
                      "kind",
                      "salePriceAmountMinor"
                    ],
                    "type": "object"
                  }
                ]
              },
              "createdAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "details": {
                "additionalProperties": false,
                "properties": {
                  "bathrooms": {
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "bedrooms": {
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "furnished": {
                    "type": "boolean"
                  },
                  "rooms": {
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "usableSurfaceSquareMeters": {
                    "exclusiveMinimum": 0,
                    "type": "number"
                  }
                },
                "type": "object"
              },
              "location": {
                "additionalProperties": false,
                "properties": {
                  "addressLine": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "city": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "country": {
                    "pattern": "^[A-Z]{2}$",
                    "type": "string"
                  },
                  "district": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "addressLine",
                  "city",
                  "country",
                  "district"
                ],
                "type": "object"
              },
              "photos": {
                "items": {
                  "$ref": "#/components/schemas/PropertyPhoto"
                },
                "type": "array"
              },
              "primaryPhoto": {
                "$ref": "#/components/schemas/PropertyPhoto"
              },
              "propertyId": {
                "$ref": "#/components/schemas/PropertyId"
              },
              "propertyType": {
                "enum": [
                  "APARTMENT",
                  "COMMERCIAL",
                  "HOUSE",
                  "LAND",
                  "OTHER"
                ],
                "type": "string"
              },
              "publishedAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "type": "string"
              },
              "status": {
                "const": "PUBLISHED",
                "type": "string"
              },
              "structuralRole": {
                "enum": [
                  "COMPOSITE",
                  "STANDALONE",
                  "UNIT"
                ],
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "transactionType": {
                "enum": [
                  "LONG_TERM_RENTAL",
                  "SALE",
                  "SHORT_TERM_RENTAL"
                ],
                "type": "string"
              },
              "updatedAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "type": "string"
              }
            },
            "required": [
              "createdAt",
              "location",
              "photos",
              "propertyId",
              "propertyType",
              "publishedAt",
              "status",
              "structuralRole",
              "title",
              "transactionType",
              "updatedAt"
            ],
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned MonPiole HTTP transport contracts",
    "title": "MonPiole Public API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/properties": {
      "get": {
        "operationId": "listProperties",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "DRAFT",
                "PUBLISHED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "enum": [
                "APARTMENT",
                "COMMERCIAL",
                "HOUSE",
                "LAND",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPortfolioResponse"
                }
              }
            },
            "description": "A stable page from the tenant-owned Property portfolio",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid discovery query or cursor"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden Property authority"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the authenticated tenant's private Property portfolio",
        "tags": [
          "Properties"
        ]
      },
      "post": {
        "operationId": "createProperty",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Draft property created",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden property authority"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft property",
        "tags": [
          "Properties"
        ]
      }
    },
...
  },
  "servers": [],
...
}